feat(gateway): 新增路由推送历史查询和菜单字段调整
路由推送历史:
- RouteController: 新增 getInstancePushHistory 接口
- RoutePushService: 新增接口定义
- RoutePushServiceImpl: 实现推送历史查询逻辑

菜单字段调整:
- AddSysMenuReq/UpdateSysMenuReq: 字段顺序调整
- SysMenuServiceImpl: 代码优化

+++ blink-gateway/gateway-admin/src/main/java/com/blink/gateway/base/dto/req/AddSysMenuReq.java
@@ -37,11 +37,11 @@ public class AddSysMenuReq implements Serializable {
 
 
     /**
-     * 菜单类型
+     * 菜单类型 1-目录 2-页面 3-按钮
      */
     @NotNull(message = BaseErrCodeConstant.PARAMETER_NOT_NULL)
     @FieldConstraint(name="flag1",message = BaseErrCodeConstant.PARAMETER_OUT_RANGE)
-    private Integer type;
+    private Byte type;
 
 
     /**
